requirements.txt
# Python 3.11

torch==2.10.0
torchvision==0.25.0
torchaudio==2.10.0

numpy==1.26.4
opencv-python==4.9.0.80
pillow==11.0.0

scikit-learn==1.4.2
scipy==1.10.1
scikit-image==0.24.0

pandas==2.2.3
tqdm==4.67.1

timm==0.9.16
segmentation_models_pytorch==0.5.0
transformers==4.57.1

tiatoolbox==1.6.0
openslide-python==1.3.1